- Merged all CoreData functionalities into the AppDelegate - Depreciated the other locations of CoreData storage
