+++ Vehicle Service Tracker/Depreciated CoreData Storage/Vehicle+CoreDataClass.swift
@@ -6,10 +6,10 @@
 //
 //
 
-import Foundation
-import CoreData
-
-@objc(Vehicle)
-public class Vehicle: NSManagedObject {
-
-}
+//import Foundation
+//import CoreData
+//
+//@objc(Vehicle)
+//public class Vehicle: NSManagedObject {
+//
+//}
