--- Vehicle Service Tracker/SaveData/PersistenceService.swift
@@ -1,65 +1,0 @@
-//
-//  PersistenceService.swift
-//  savingData for Vehicle Service Tracker
-//
-//  Created by Alan Crisologo on 2/12/22.
-//
-
-import Foundation
-import CoreData
-
-class PersistenceService {
-    // MARK: - Core Data stack
-
-    private init(){}
-    
-    static var context: NSManagedObjectContext {
-        return persistentContainer.viewContext
-    }
-    
-    static var persistentContainer: NSPersistentContainer = {
-        /*
-         The persistent container for the application. This implementation
-         creates and returns a container, having loaded the store for the
-         application to it. This property is optional since there are legitimate
-         error conditions that could cause the creation of the store to fail.
-        */
-        
-                                            //UPDATE THIS VARIABLE BELOW TO YOUR PROJECT NAME
-        let container = NSPersistentContainer(name: "Vehicle Service Tracker")
-        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-            if let error = error as NSError? {
-                // Replace this implementation with code to handle the error appropriately.
-                // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
-                 
-                /*
-                 Typical reasons for an error here include:
-                 * The parent directory does not exist, cannot be created, or disallows writing.
-                 * The persistent store is not accessible, due to permissions or data protection when the device is locked.
-                 * The device is out of space.
-                 * The store could not be migrated to the current model version.
-                 Check the error message to determine what the actual problem was.
-                 */
-                fatalError("Unresolved error \(error), \(error.userInfo)")
-            }
-        })
-        return container
-    }()
-
-    // MARK: - Core Data Saving support
-
-    static func saveContext () {
-        let context = persistentContainer.viewContext
-        if context.hasChanges {
-            do {
-                try context.save()
-                print("Saved!")
-            } catch {
-                // Replace this implementation with code to handle the error appropriately.
-                // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
-                let nserror = error as NSError
-                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
-            }
-        }
-    }
-}
