+++ Vehicle Service Tracker/SaveData/Vehicle Service Tracker.xcdatamodeld/savingData.xcdatamodel/contents
@@ -1,18 +1,14 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="20086" systemVersion="21D62" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="Vehicle" representedClassName="Vehicle" syncable="YES">
-        <attribute name="carPic" optional="YES" attributeType="Binary"/>
+    <entity name="Vimage" representedClassName="Vimage" syncable="YES" codeGenerationType="class">
+        <attribute name="img" optional="YES" attributeType="Binary" allowsExternalBinaryDataStorage="YES"/>
         <attribute name="lastServiceDate" optional="YES" attributeType="String"/>
         <attribute name="makeModel" optional="YES" attributeType="String"/>
         <attribute name="purchaseDate" optional="YES" attributeType="String"/>
         <attribute name="totalMiles" optional="YES" attributeType="String"/>
         <attribute name="year" optional="YES" attributeType="String"/>
     </entity>
-    <entity name="Vimage" representedClassName="Vimage" syncable="YES" codeGenerationType="class">
-        <attribute name="img" optional="YES" attributeType="Binary" allowsExternalBinaryDataStorage="YES"/>
-    </entity>
     <elements>
-        <element name="Vehicle" positionX="-63" positionY="-18" width="128" height="119"/>
-        <element name="Vimage" positionX="-54" positionY="27" width="128" height="44"/>
+        <element name="Vimage" positionX="-54" positionY="27" width="128" height="119"/>
     </elements>
 </model>
