+++ Vehicle Service Tracker/Vehicles Tab/CustomCell.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 class CustomCell: UITableViewCell {
-
+    
     @IBOutlet weak var carView: UIView!
     @IBOutlet weak var carLabel: UILabel!
     @IBOutlet weak var carYearLabel: UILabel!
@@ -18,11 +18,11 @@ class CustomCell: UITableViewCell {
         super.awakeFromNib()
         // Initialization code
     }
-
+    
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
-
+        
         // Configure the view for the selected state
     }
-
+    
 }
